gh-108303: Move `double_const` to `test_import` where it belongs
